Pull the centroid index reading into a separate class
