+++ server/src/main/java/org/elasticsearch/index/codec/vectors/diskbbq/next/FlatCentroidIndex.java
@@ -9,8 +9,10 @@
 
 package org.elasticsearch.index.codec.vectors.diskbbq.next;
 
+import org.apache.lucene.index.FieldInfo;
 import org.apache.lucene.index.FloatVectorValues;
 import org.apache.lucene.index.KnnVectorValues;
+import org.apache.lucene.index.VectorSimilarityFunction;
 import org.apache.lucene.search.AcceptDocs;
 import org.apache.lucene.search.ConjunctionUtils;
 import org.apache.lucene.search.DocIdSetIterator;
@@ -20,6 +22,9 @@ import org.apache.lucene.util.LongValues;
 import org.apache.lucene.util.packed.DirectReader;
 import org.apache.lucene.util.packed.DirectWriter;
 import org.elasticsearch.index.codec.vectors.OptimizedScalarQuantizer;
+import org.elasticsearch.index.codec.vectors.cluster.NeighborQueue;
+import org.elasticsearch.index.codec.vectors.diskbbq.CentroidIterator;
+import org.elasticsearch.index.codec.vectors.diskbbq.PostingMetadata;
 import org.elasticsearch.search.vectors.ESAcceptDocs;
 import org.elasticsearch.simdvec.ES92Int7VectorsScorer;
 import org.elasticsearch.simdvec.ESVectorUtil;
@@ -27,26 +32,56 @@ import org.elasticsearch.simdvec.ESVectorUtil;
 import java.io.IOException;
 import java.util.List;
 
+import static org.elasticsearch.index.codec.vectors.diskbbq.PostingMetadata.NO_ORDINAL;
+
 class FlatCentroidIndex implements CentroidIndex {
 
-    FlatCentroidIndex() {
-        final ESNextDiskBBQVectorsReader.NextFieldEntry fieldEntry = fields.get(fieldInfo.number);
-        // build optmization filters if possible
-        final FixedBitSet acceptCentroids = getCentroidFilter(centroids, numCentroids, values, acceptDocs, approximateCost);
-        final int numParents = centroids.readVInt();
-        final FixedBitSet acceptParents = getParentCentroidFilter(centroids, numParents, numCentroids, acceptDocs, fieldEntry.numSlices);
+    private final FieldInfo fieldInfo;
+    private final ESNextDiskBBQVectorsReader.NextFieldEntry fieldEntry;
+    private final int numCentroids;
+    private final IndexInput centroids;
+    private final float visitRatio;
+    private final FixedBitSet acceptCentroids;
+    private final int numParents;
+    private final FixedBitSet acceptParents;
+    private final int bulkSize;
+    private final byte[] quantized;
+    private final OptimizedScalarQuantizer.QuantizationResult queryParams;
+
+    FlatCentroidIndex(
+        FieldInfo fieldInfo,
+        ESNextDiskBBQVectorsReader.NextFieldEntry fieldEntry,
+        int numCentroids,
+        IndexInput centroids,
+        float[] targetQuery,
+        AcceptDocs acceptDocs,
+        float approximateCost,
+        FloatVectorValues values,
+        float visitRatio
+    ) throws IOException {
+        this.fieldInfo = fieldInfo;
+        this.fieldEntry = fieldEntry;
+        this.numCentroids = numCentroids;
+        this.centroids = centroids;
+        this.visitRatio = visitRatio;
+
+        // build optimization filters if possible
+        acceptCentroids = getCentroidFilter(centroids, numCentroids, values, acceptDocs, approximateCost);
+        numParents = centroids.readVInt();
+        acceptParents = getParentCentroidFilter(centroids, numParents, numCentroids, acceptDocs, fieldEntry.numSlices);
+
         // build centroid search helpers
-        final int bulkSize = fieldEntry.getBulkSize();
-        final OptimizedScalarQuantizer scalarQuantizer = new OptimizedScalarQuantizer(fieldInfo.getVectorSimilarityFunction());
-        final int[] scratch = new int[targetQuery.length];
-        final OptimizedScalarQuantizer.QuantizationResult queryParams = scalarQuantizer.scalarQuantize(
+        bulkSize = fieldEntry.getBulkSize();
+        OptimizedScalarQuantizer scalarQuantizer = new OptimizedScalarQuantizer(fieldInfo.getVectorSimilarityFunction());
+        int[] scratch = new int[targetQuery.length];
+        queryParams = scalarQuantizer.scalarQuantize(
             targetQuery,
             new float[targetQuery.length],
             scratch,
             (byte) 7,
             fieldEntry.globalCentroid()
         );
-        final byte[] quantized = new byte[targetQuery.length];
+        quantized = new byte[targetQuery.length];
         for (int i = 0; i < quantized.length; i++) {
             quantized[i] = (byte) scratch[i];
         }
@@ -127,14 +162,13 @@ class FlatCentroidIndex implements CentroidIndex {
     }
 
     @Override
-    public CentroidIterator getIterator() {
+    public CentroidIterator getIterator() throws IOException {
         final ES92Int7VectorsScorer scorer = ESVectorUtil.getES92Int7VectorsScorer(centroids, fieldInfo.getVectorDimension(), bulkSize);
         // build iterator
-        CentroidIterator centroidIterator;
         if (numParents > 0) {
             // equivalent to (float) centroidsPerParentCluster / 2
             float centroidOversampling = (float) fieldEntry.numCentroids() / (2 * numParents);
-            centroidIterator = getCentroidIteratorWithParents(
+            return getCentroidIteratorWithParents(
                 fieldInfo,
                 centroids,
                 numParents,
@@ -152,7 +186,7 @@ class FlatCentroidIndex implements CentroidIndex {
             if (acceptCentroids != null && acceptParents != null) {
                 acceptCentroids.and(acceptParents);
             }
-            centroidIterator = getCentroidIteratorNoParent(
+            return getCentroidIteratorNoParent(
                 fieldInfo,
                 centroids,
                 numCentroids,
@@ -163,6 +197,303 @@ class FlatCentroidIndex implements CentroidIndex {
                 acceptCentroids != null ? acceptCentroids : acceptParents,
                 bulkSize
             );
+        }
+    }
+
+    private static CentroidIterator getCentroidIteratorNoParent(
+        FieldInfo fieldInfo,
+        IndexInput centroids,
+        int numCentroids,
+        ES92Int7VectorsScorer scorer,
+        byte[] quantizeQuery,
+        OptimizedScalarQuantizer.QuantizationResult queryParams,
+        float globalCentroidDp,
+        FixedBitSet acceptCentroids,
+        int bulkSize
+    ) throws IOException {
+        final NeighborQueue neighborQueue = new NeighborQueue(numCentroids, true);
+        final long centroidQuantizeSize = fieldInfo.getVectorDimension() + 3 * Float.BYTES + Integer.BYTES;
+        score(
+            neighborQueue,
+            numCentroids,
+            0,
+            scorer,
+            centroids,
+            centroidQuantizeSize,
+            quantizeQuery,
+            queryParams,
+            globalCentroidDp,
+            fieldInfo.getVectorSimilarityFunction(),
+            new float[bulkSize],
+            acceptCentroids,
+            bulkSize
+        );
+        long offset = centroids.getFilePointer();
+        return new CentroidIterator() {
+            @Override
+            public boolean hasNext() {
+                return neighborQueue.size() > 0;
+            }
+
+            @Override
+            public PostingMetadata nextPosting() throws IOException {
+                long centroidOrdinalAndScore = neighborQueue.popRaw();
+                int centroidOrd = neighborQueue.decodeNodeId(centroidOrdinalAndScore);
+                float score = neighborQueue.decodeScore(centroidOrdinalAndScore);
+                centroids.seek(offset + (long) Long.BYTES * 2 * centroidOrd);
+                long postingListOffset = centroids.readLong();
+                long postingListLength = centroids.readLong();
+                // NO_ORDINAL indicates that the global centroid should be used for query quantization
+                return new PostingMetadata(postingListOffset, postingListLength, NO_ORDINAL, score);
+            }
+        };
+    }
+
+    private static CentroidIterator getCentroidIteratorWithParents(
+        FieldInfo fieldInfo,
+        IndexInput centroids,
+        int numParents,
+        int numCentroids,
+        ES92Int7VectorsScorer scorer,
+        byte[] quantizeQuery,
+        OptimizedScalarQuantizer.QuantizationResult queryParams,
+        float globalCentroidDp,
+        float centroidRatio,
+        FixedBitSet acceptParents,
+        FixedBitSet acceptCentroids,
+        int bulkSize
+    ) throws IOException {
+        // build the three queues we are going to use
+        final long rawParentSize = (long) fieldInfo.getVectorDimension() * Float.BYTES;
+        final long centroidQuantizeSize = fieldInfo.getVectorDimension() + 3 * Float.BYTES + Integer.BYTES;
+        final NeighborQueue parentsQueue = new NeighborQueue(numParents, true);
+        final int maxChildrenSize = centroids.readVInt();
+        final NeighborQueue currentParentQueue = new NeighborQueue(maxChildrenSize, true);
+        final int bufferSize = (int) Math.clamp(centroidRatio * numCentroids, 1, numCentroids);
+        final int numCentroidsFiltered = acceptCentroids == null ? numCentroids : acceptCentroids.cardinality();
+        if (numCentroidsFiltered == 0) {
+            // TODO maybe this makes CentroidIterator polymorphic?
+            return new CentroidIterator() {
+                @Override
+                public boolean hasNext() {
+                    return false;
+                }
+
+                @Override
+                public PostingMetadata nextPosting() {
+                    return null;
+                }
+            };
+        }
+        final float[] scores = new float[bulkSize];
+        final NeighborQueue neighborQueue;
+        if (acceptCentroids != null && numCentroidsFiltered <= bufferSize) {
+            // we are collecting every non-filter centroid, therefore we do not need to score the
+            // parents. We give each of them the same score.
+            neighborQueue = new NeighborQueue(numCentroidsFiltered, true);
+            for (int i = 0; i < numParents; i++) {
+                if (acceptParents == null || acceptParents.get(i)) {
+                    parentsQueue.add(i, 0.5f);
+                }
+            }
+            centroids.skipBytes((centroidQuantizeSize + rawParentSize) * numParents);
+        } else {
+            neighborQueue = new NeighborQueue(bufferSize, true);
+            // score the parents
+            centroids.skipBytes(rawParentSize * numParents);
+            score(
+                parentsQueue,
+                numParents,
+                0,
+                scorer,
+                centroids,
+                centroidQuantizeSize,
+                quantizeQuery,
+                queryParams,
+                globalCentroidDp,
+                fieldInfo.getVectorSimilarityFunction(),
+                scores,
+                acceptParents,
+                bulkSize
+            );
+        }
+
+        final long offset = centroids.getFilePointer();
+        final long childrenOffset = offset + (long) Long.BYTES * numParents;
+        // populate the children's queue by reading parents one by one
+        while (parentsQueue.size() > 0 && neighborQueue.size() < bufferSize) {
+            final int pop = parentsQueue.pop();
+            populateOneChildrenGroup(
+                currentParentQueue,
+                centroids,
+                offset + 2L * Integer.BYTES * pop,
+                childrenOffset,
+                centroidQuantizeSize,
+                fieldInfo,
+                scorer,
+                quantizeQuery,
+                queryParams,
+                globalCentroidDp,
+                scores,
+                acceptCentroids,
+                bulkSize
+            );
+            while (currentParentQueue.size() > 0 && neighborQueue.size() < bufferSize) {
+                final float score = currentParentQueue.topScore();
+                final int children = currentParentQueue.pop();
+                neighborQueue.add(children, score);
+            }
+        }
+        final long childrenFileOffsets = childrenOffset + centroidQuantizeSize * numCentroids;
+        return new CentroidIterator() {
+
+            @Override
+            public boolean hasNext() {
+                return neighborQueue.size() > 0;
+            }
+
+            @Override
+            public PostingMetadata nextPosting() throws IOException {
+                long centroidOrdinalAndScore = nextCentroid();
+                int centroidOrdinal = neighborQueue.decodeNodeId(centroidOrdinalAndScore);
+                float score = neighborQueue.decodeScore(centroidOrdinalAndScore);
+                centroids.seek(childrenFileOffsets + (long) (Long.BYTES * 2 + Integer.BYTES) * centroidOrdinal);
+                long postingListOffset = centroids.readLong();
+                long postingListLength = centroids.readLong();
+                int parentOrd = centroids.readInt();
+                return new PostingMetadata(postingListOffset, postingListLength, parentOrd, score);
+            }
+
+            private long nextCentroid() throws IOException {
+                if (currentParentQueue.size() > 0) {
+                    // return next centroid and maybe add a children from the current parent queue
+                    return neighborQueue.popRawAndAddRaw(currentParentQueue.popRaw());
+                } else if (parentsQueue.size() > 0) {
+                    // current parent queue is empty, populate it again with the next parent
+                    int pop = parentsQueue.pop();
+                    populateOneChildrenGroup(
+                        currentParentQueue,
+                        centroids,
+                        offset + 2L * Integer.BYTES * pop,
+                        childrenOffset,
+                        centroidQuantizeSize,
+                        fieldInfo,
+                        scorer,
+                        quantizeQuery,
+                        queryParams,
+                        globalCentroidDp,
+                        scores,
+                        acceptCentroids,
+                        bulkSize
+                    );
+                    return nextCentroid();
+                } else {
+                    return neighborQueue.popRaw();
+                }
+            }
+        };
+    }
+
+    private static void populateOneChildrenGroup(
+        NeighborQueue neighborQueue,
+        IndexInput centroids,
+        long parentOffset,
+        long childrenOffset,
+        long centroidQuantizeSize,
+        FieldInfo fieldInfo,
+        ES92Int7VectorsScorer scorer,
+        byte[] quantizeQuery,
+        OptimizedScalarQuantizer.QuantizationResult queryParams,
+        float globalCentroidDp,
+        float[] scores,
+        FixedBitSet acceptCentroids,
+        int bulkSize
+    ) throws IOException {
+        centroids.seek(parentOffset);
+        int childrenOrdinal = centroids.readInt();
+        int numChildren = centroids.readInt();
+        centroids.seek(childrenOffset + centroidQuantizeSize * childrenOrdinal);
+        score(
+            neighborQueue,
+            numChildren,
+            childrenOrdinal,
+            scorer,
+            centroids,
+            centroidQuantizeSize,
+            quantizeQuery,
+            queryParams,
+            globalCentroidDp,
+            fieldInfo.getVectorSimilarityFunction(),
+            scores,
+            acceptCentroids,
+            bulkSize
+        );
+    }
+
+    private static void score(
+        NeighborQueue neighborQueue,
+        int size,
+        int scoresOffset,
+        ES92Int7VectorsScorer scorer,
+        IndexInput centroids,
+        long centroidQuantizeSize,
+        byte[] quantizeQuery,
+        OptimizedScalarQuantizer.QuantizationResult queryCorrections,
+        float centroidDp,
+        VectorSimilarityFunction similarityFunction,
+        float[] scores,
+        FixedBitSet acceptCentroids,
+        int bulkSize
+    ) throws IOException {
+        int limit = size - bulkSize + 1;
+        int i = 0;
+        for (; i < limit; i += bulkSize) {
+            if (acceptCentroids == null || acceptCentroids.cardinality(scoresOffset + i, scoresOffset + i + bulkSize) > 0) {
+                scorer.scoreBulk(
+                    quantizeQuery,
+                    queryCorrections.lowerInterval(),
+                    queryCorrections.upperInterval(),
+                    queryCorrections.quantizedComponentSum(),
+                    queryCorrections.additionalCorrection(),
+                    similarityFunction,
+                    centroidDp,
+                    scores,
+                    bulkSize
+                );
+                for (int j = 0; j < bulkSize; j++) {
+                    int centroidOrd = scoresOffset + i + j;
+                    if (acceptCentroids == null || acceptCentroids.get(centroidOrd)) {
+                        neighborQueue.add(centroidOrd, scores[j]);
+                    }
+                }
+            } else {
+                centroids.skipBytes(bulkSize * centroidQuantizeSize);
+            }
+        }
+
+        int tailBulkSize = size - i;
+        if (tailBulkSize > 0) {
+            if (acceptCentroids == null || acceptCentroids.cardinality(scoresOffset + i, scoresOffset + i + tailBulkSize) > 0) {
+                scorer.scoreBulk(
+                    quantizeQuery,
+                    queryCorrections.lowerInterval(),
+                    queryCorrections.upperInterval(),
+                    queryCorrections.quantizedComponentSum(),
+                    queryCorrections.additionalCorrection(),
+                    similarityFunction,
+                    centroidDp,
+                    scores,
+                    tailBulkSize
+                );
+                for (int j = 0; j < tailBulkSize; j++) {
+                    int centroidOrd = scoresOffset + i + j;
+                    if (acceptCentroids == null || acceptCentroids.get(centroidOrd)) {
+                        neighborQueue.add(centroidOrd, scores[j]);
+                    }
+                }
+            } else {
+                centroids.skipBytes(tailBulkSize * centroidQuantizeSize);
+            }
         }
     }
 }
